refactored controllers and routes.  Enable frontend
